# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mall DRC API",
        "version": "1.0"
    },
    "paths": {
        "/categorie": {
            "get": {
                "tags": [
                    "categorie"
                ],
                "summary": "Affiche Toutes les categories se trouvant dans la base de données",
                "operationId": "3098b3025960cb15ab87d997c9754f88",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "categorie"
                ],
                "summary": "Enregistrer une categorie",
                "operationId": "saveCategorie",
                "responses": {
                    "201": {
                        "description": "Enregistremnt avec success"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/categorie/{idCategorie}": {
            "get": {
                "tags": [
                    "categorie"
                ],
                "summary": " Affiche Toutes les categories ayant l'Id recherché",
                "operationId": "4730be719d42fa8083a68f9c43a58aa9",
                "parameters": [
                    {
                        "name": "idCategorie",
                        "in": "path",
                        "description": "idCategorie designe l'Id recherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "12"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "visualisation du categorie recherché"
                    },
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "le categorie n'existe pas"
                    }
                }
            }
        },
        "/categorie/{categorie}": {
            "delete": {
                "tags": [
                    "categorie"
                ],
                "summary": "supprimes le categorie ayant Id specifier à l'argument",
                "operationId": "destroyCategorie",
                "parameters": [
                    {
                        "name": "categorie",
                        "in": "path",
                        "description": "categorie est l'Id du prosui récherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "12"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "le produit n'existe pas"
                    }
                }
            }
        },
        "/categorie/search/libelleCategorie": {
            "get": {
                "tags": [
                    "categorie"
                ],
                "summary": " Affiche Toutes les categorie contenant le nom rechercher dans son libellé",
                "operationId": "searchCategorie",
                "parameters": [
                    {
                        "name": "produit",
                        "in": "path",
                        "description": "libelleCategorie designe toutes les categories contenant le nom rechercher son libellé",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "chemise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "nom": {
                                                        "type": "string",
                                                        "example": "ballon"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/client/{client}": {
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "supprimes les client ayant Id specifier à l'argument",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "client est l'Id du client récherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "client n'existe pas"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:client",
                            "read:client"
                        ]
                    }
                ]
            }
        },
...
        "/marchands/{marchand}": {
            "get": {
                "tags": [
                    "marchand"
                ],
                "summary": "affiche un marchand spécifique ayant Id specifier à l'argument",
                "operationId": "showMarchand",
                "parameters": [
                    {
                        "name": "marchand",
                        "in": "path",
                        "description": "marchand est l'Id du marchand récherché",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "visualisation du produit recherché"
                    },
                    "400": {
                        "description": "Id Invalide"
                    },
                    "404": {
                        "description": "le produit n'existe pas"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:produit",
                            "read:produit"
                        ]
                    }
                ]
            }
        },
...
    }
}